roots: re-structured roots for future functions
I've restructured how roots are parsed in anticipation of future
function declarations & definitions so that function signatures
are parsed first before the defintions so that functions calls are 'hoisted'.

I'll also have to implement a future function signature table and also state to keep track of and prevent recursion (will cause errors & infinite loop due to the lack of a stack).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>