parser: support for function definitions implemented
I've implemented parsing for function definitions; with parameters, return
types, bodies- the whole shebang.

So far, duplicate function definitions are *allowed* and invalid return
types are also *valid* in the parser as they should be handled by the type
system.

Function signatures should be parsed first in the type system to allow
for 'hoisting' for function calls (before the parsing of both the main
block and also the function bodies).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>