type system: support for function definitions
I've implemented support for function definitions in the type system
(through the wrapping of the function body in types and verifying both
the return and parameter types).

I've also changed the parsing rules so that return types must be explicitly
stated in the function definition to make things simpler as I don't want to
have to infer the type based on a return expression.

return, break and continue blocks will likely *not* be implemented as they
far too complicated though a 'guard' statement may be implemented in future
to replace the functionality of it. (a clean way to wrap the rest of a scope
in an else statement).

Like so:
```scrapile
if (some_condition)
  something
guard

// everything else is in an implicit else block to prevent nesting
```

There is no checking so far for recursion nor nested lists (both invalid).

There is also now a function-table for verifying function calls.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>