function support complete
I've fully implemented support for functions in the underlying scratch
target which signals the completeness of functions in this initial langauge
implementation.

Lists are supported as parameters and arguments and function calls are even
nestable, suprisingly. Lists are also now support as return types from
functions through the use of temporary binds (same logic as setting a
variable to a list).

Recursion is currently not valid due to a lack of a stack and any attempt
in doing so will break the code. There is no hard compiler error for it
currently as tracing function calls prove to be difficult.

Support for guard clauses (return alternatives), maths shortenings
(like `+=` or `*=`) and also further string operations (like indexing)
will be next to implement before the project is rewritten.

Some code for dealing with and translating lists have also been changed to be reproducable on a second run (through the wiping of the mutated list beforehand).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>